hacking out a solution to changing map color var
Rather than forever just showing sum of VMT, mechanism to change the
variables.

First, need to do the iterative summing in Elm, not JS.

Figured that out, and have the basic framework in place to read required
data from the model, but the code won't compile.

Save and fix later