seems to work
left a buncha commented stuff in there.  nasty looking commit.  Anyway,
not so sure this is good, but it looks okay and isn't crashing, so it
must be okay, right?  Also, the running sum is right, and the max values
are the same as the old version running on metis, so that's good